Clarify Sorting(s)
It didn't really make sense why we had both Sorting and Sortings. It
turns out Sortings was just a factory for picking a Sorting, so I
converted its functions to a class method on Sorting. I then renamed
it to keep the new conceptual space for it separate.

Not super thrilled with this -- there's ALMOST some kind of paralellism
going on with the Search, Filter, and Ordering aspects on Notice, but
not quite. Also I've wondered for a long time if those want to be moved
outside of Notice, to slim down that model a bit -- the logic was
written to generalize across Notice and Entity but only Notice uses
it -- but I haven't quite convinced myself that's right, or how to go
about it.